+++ src/TitleForm/TitleForm.test.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { useHistory } from 'react-router';
 import { Form } from 'react-final-form';
-import { render } from '@testing-library/react';
-import user from '@testing-library/user-event';
+import { render } from '@folio/jest-config-stripes/testing-library/react';
+import user from '@folio/jest-config-stripes/testing-library/user-event';
 
 import {
   HasCommand,
@@ -65,10 +65,10 @@ describe('TitleForm', () => {
   });
 
   describe('Close form', () => {
-    it('should close Title form', () => {
+    it('should close Title form', async () => {
       const { getByText } = renderTitleForm();
 
-      user.click(getByText('stripes-acq-components.FormFooter.cancel'));
+      await user.click(getByText('stripes-acq-components.FormFooter.cancel'));
 
       expect(defaultProps.onCancel).toHaveBeenCalled();
     });