+++ src/TitleReceive/TitleReceiveContainer.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { act, render, cleanup, screen } from '@testing-library/react';
+import { act, render, cleanup, screen } from '@folio/jest-config-stripes/testing-library/react';
 import { IntlProvider } from 'react-intl';
 import { MemoryRouter } from 'react-router-dom';