+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Change history for ui-receiving
 
-## (IN PROGRESS)
+## 4.0.0 (IN PROGRESS)
 
 * Unpin `@rehooks/local-storage` now that it's no longer broken. Refs UIREC-259.
 * Also support `circulation` `14.0`. Refs UIREC-270.
 * Receiving an item with an open request, the warning dialogue shows "undefined" instead of title. Refs UIREC-278.
+* Upgrade `Node.js` to `18` version in GitHub Actions. Refs UIREC-281.
+* *BREAKING* Upgrade React to v18. Refs UIREC-280.
 
 ## [3.0.0](https://github.com/folio-org/ui-receiving/tree/v3.0.0) (2023-02-22)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v2.3.1...v3.0.0)